# docassistant-openapi.yaml
openapi: 3.1.0
info:
  title: DocAssistant API
  description: API for searching and summarizing documents
  version: 1.0.0
  contact:
    name: DocAssistant Team
  license:
    name: MIT
    
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Local development server

paths:
  /search:
    post:
      summary: Search for documents
      description: Search for documents using various criteria including text, dates, and metadata
      operationId: searchDocuments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentSearchRequest'
      responses:
        '200':
          description: Successful search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSearchResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /summary:
    post:
      summary: Generate summaries for documents
      description: Generate AI-powered summaries for specified documents
      operationId: generateSummaries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentSummaryRequest'
      responses:
        '200':
          description: Successfully generated summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSummaryResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    DocumentSearchRequest:
      type: object
      properties:
        search_text:
          type: string
          description: Text to search for in document title or abstract
          example: "renewable energy"
        start_date:
          type: string
          format: date
          description: Start date for document search range
          example: "2023-01-01"
        end_date:
          type: string
          format: date
          description: End date for document search range
          example: "2023-12-31"
        doc_type:
          type: string
          description: Exact match on type of document to search for
          example: "research_paper"
        language:
          type: string
          description: Exact match on language of the documents
          example: "en"
        country:
          type: string
          description: Exact match on country associated with the documents
          example: "USA"

    DocumentSearchResponse:
      type: object
      required:
        - results
        - result_count
        - search_time_ms
      properties:
        results:
          type: array
          description: List of matching documents
          items:
            $ref: '#/components/schemas/DocumentResult'
        result_count:
          type: integer
          description: Total number of matching documents
          example: 42
        search_time_ms:
          type: integer
          description: Time taken to perform the search in milliseconds
          example: 123

    DocumentResult:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: string
          description: Unique document identifier
          example: "doc_123"
        title:
          type: string
          description: Document title
          example: "Analysis of Renewable Energy Trends"
        abstract:
          type: string
          description: Document abstract or summary
          example: "This paper analyzes global trends in renewable energy adoption..."
        doc_date:
          type: string
          format: date
          description: Document publication date
          example: "2023-03-15"
        doc_type:
          type: string
          description: Type of document
          example: "research_paper"
        language:
          type: string
          description: Document language
          example: "en"
        country:
          type: string
          description: Country associated with the document
          example: "USA"

    DocumentSummaryRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          description: List of document IDs to summarize
          items:
            type: string
          example: ["doc_123", "doc_456"]
        model:
          type: string
          description: AI model to use for summarization
          example: "gpt-4"
        modelOptions:
          type: object
          description: Additional options for the AI model
          additionalProperties:
            type: string
          example:
            max_length: "500"
            temperature: "0.7"

    DocumentSummaryResponse:
      type: object
      required:
        - summary_text
        - summary_time_ms
      properties:
        summary_text:
          type: string
          description: Generated summary text
          example: "The documents discuss advancements in renewable energy..."
        summary_time_ms:
          type: integer
          description: Time taken to generate the summary in milliseconds
          example: 2345

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "INVALID_PARAMETER"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid date format for start_date"
        details:
          type: object
          description: Additional error details
          additionalProperties: true

tags:
  - name: search
    description: Document search operations
  - name: summary
    description: Document summarization operations

security: []